+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [9.4.3] - (18.11.2024)
+## [9.5.0] - (19.11.2024)
+
+## Changes
+
+- added possibility to select compare mode for power and energy level sensors
+- config schema updated
+- readme updated
+- cleanup
+
+## [9.4.4] - (18.11.2024)
 
 ## Changes